examples/langchain-bot/requirements.txt
google-cloud-aiplatform==1.39.0
langchain==0.0.310
# needed for retrieving logs (and writing too if needed)
google-cloud-logging>=3.0.0
Flask
# core needed for custom agents
# langchain-core

# prettyprinter==0.18.0
# wikipedia==1.4.0
# chromadb==0.3.26
# chromadb
# tiktoken==0.5.1
# tabulate==0.9.0
# sqlalchemy-bigquery==1.8.0
# google-cloud-bigquery==3.11.4
# pillow==9.4.0
# embeddings's requirements
# google-cloud-storage
# google-cloud-bigquery[pandas]
#pyarrow is needed for db-dtypes (See README for troubleshooting)
# pyarrow==3.0.0
# db-dtypes
# pymupdf
# seaborn